+++ src/@types/bottom-sheet/index.ts
@@ -100,6 +100,18 @@ export interface SmartBottomSheetProps {
    * @default windowHeight * 0.4
    */
   closeDragHeight?: number;
+  /**
+   * 是否锁定背景滚动
+   * @version 2.9.0
+   * @default true
+   */
+  lockScroll?: boolean;
+  /**
+   * 是否锁定最大拖拽高度，设置为 `true` 时，拖拽过程中面板高度不会超过 `max-drag-height`
+   * @version 2.9.0
+   * @default false
+   */
+  lockMaxDrag?: boolean;
 }
 
 export interface SmartBottomSheetEvents extends SmartTransitionEvents {