+++ src/@types/config-provider/theme-vars.ts
@@ -301,7 +301,6 @@ type CalendarThemeVars =
   | 'calendarInfoFontSize'
   | 'calendarInfoLineHeight'
   | 'calendarSelectedDayBackgroundColor'
-  | 'calendarDayDisabledColor'
   | 'calendarConfirmButtonHeight'
   | 'calendarConfirmButtonMargin'
   | 'calendarConfirmButtonLineHeight'
@@ -780,6 +779,7 @@ type CascaderThemeVars =
   | 'cascaderSelectedIconSize'
   | 'cascaderTabsHeight'
   | 'cascaderActiveColor'
+  | 'cascaderActiveBackgroundColor'
   | 'cascaderOptionsHeight'
   | 'cascaderOptionDisabledColor'
   | 'cascaderTabColor'
@@ -798,7 +798,12 @@ type CustomKeyboardThemeVars =
   | 'customKeyboardPopupTextColor'
   | 'customKeyboardPopupHoverColor';
 
-type PopoverThemeVars = 'popoverBackgroundColor' | 'popoverBoxShadow' | 'popoverPadding' | 'popoverBorderRadius';
+type PopoverThemeVars =
+  | 'popoverBackgroundColor'
+  | 'popoverBoxShadow'
+  | 'popoverPadding'
+  | 'popoverBorderRadius'
+  | 'popoverOverlayColor';
 
 /**
  * 主题变量